requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
greenlet==3.3.0
idna==3.11
joblib==1.5.2
numpy==1.26.4
pandas==2.1.4
psycopg==3.3.2
psycopg-binary==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.4.2
scipy==1.15.3
six==1.16.0
SQLAlchemy==2.0.29
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.6.2
